Feature/40 - page for compare users' tastes

+++ users/migrations/0005_alter_user_avatar_alter_user_birthday_and_more.py
@@ -70,7 +70,7 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name="user",
             name="yandex_id",
-            field=models.IntegerField(
+            field=models.BigIntegerField(
                 blank=True, null=True, unique=True, verbose_name="идентификатор в Яндексе"
             ),
         ),
